+++ lib/test.ts
@@ -1,10 +1,30 @@
 import { Logger } from "./Logger.js";
-import { IOLevelLog } from "./LoggerInterfaces.js";
+import { format } from "util";
 
 const logger = new Logger({
   format: "pretty",
-  levelLog: [IOLevelLog.ALL],
+  levelLog: Number.POSITIVE_INFINITY,
   useColor: true,
+});
+
+logger.on("logging", (lvl, data, msg) => {
+  process.stdout.write(format.apply(null, [msg]));
+});
+
+logger.on("fatalLogging", (lvl, data, timeStamp, prefix, errors) => {
+  process.stdout.write(format.apply(null, [prefix]));
+});
+
+logger.on("levelChange", (lvl, prev) => {
+  process.stdout.write(
+    format.apply(null, ["Previous level:", prev, "Now level:", lvl])
+  );
+});
+
+logger.on("settingChange", (prev, newSetting) => {
+  process.stdout.write(
+    format.apply(null, ["Previous setting:", prev, "Now setting: ", newSetting])
+  );
 });
 
 console.time("time");
@@ -13,15 +33,8 @@ logger.info("Hello");
 
 console.timeEnd("time");
 
-// // const info = logger.info("hi");
+const childLogger = logger.child({ a: "b" }, { isChild: true });
 
-// const childLogger = logger.child(
-//   { thisIsChild: true },
-//   { thisIsLoggerOpt: true, hi: "hi" }
-// );
+childLogger.getBindingOpt().a;
 
-// console.time("time");
-
-// const info2 = childLogger.info("hi");
-
-// console.timeEnd("time");
+childLogger.loggerProps.isChild;