+++ CHANGELOG.md
@@ -1,5 +1,121 @@
 # Start From 1.6.7 version
 
+## 1.6.10
+
+- Increase performance
+
+- Remove `IOLevelLog.ALL`
+
+- Added `events.EventEmitter` with 4 events:
+
+  - `levelChange`
+  - `logging`
+  - `fatalLogging`
+  - `settingChange`
+
+- Bring back `onInit()`
+
+  ```ts
+  const logger = new Logger(
+    {
+      format: "pretty",
+      levelLog: Number.POSITIVE_INFINITY,
+      useColor: true,
+    },
+    (Logger) => {
+      Logger.info("Logger is ready");
+    }
+  );
+  ```
+
+- Changed `stacktrace.ts` to `error-stack-parser` package
+
+- `loggerProps` and `childProps` now is read-only
+
+  ```ts
+  const logger = new Logger({
+    format: "pretty",
+    levelLog: Number.POSITIVE_INFINITY,
+    useColor: true,
+  });
+
+  const childLogger = logger.child({ a: "b" }, { isChild: true });
+
+  childLogger.getBindingOpt().a; /* Nothing happend */
+  childLogger.getBindingOpt().a = "c"; /* This line will throw Error */
+
+  childLogger.loggerProps.isChild; /* Nothing happend */
+  childLogger.loggerProps.isChild = false; /* This line will throw Error */
+  ```
+
+- Edited `IOErrorStack`
+
+  ```ts
+  interface IOErrorStack {
+    filePath: string;
+    fullFilePath: string | undefined;
+    lineNumber: number | undefined;
+    lineColumm: number | undefined;
+    functionName: string | undefined;
+    methodName: string | undefined;
+    isConstructor: boolean | undefined;
+  }
+  ```
+
+- Added `getBindingOpt()` to get `childProps`
+
+  ```ts
+  const logger = new Logger({
+    format: "pretty",
+    levelLog: Number.POSITIVE_INFINITY,
+    useColor: true,
+  });
+
+  const childLogger = logger.child({ a: "b" }, { isChild: true });
+
+  console.log("Binding:", childLogger.getBindingOpt().a);
+
+  /*----------------Output----------------*/
+  /*
+  Binding: b
+  */
+  ```
+
+- Added `useExpressLogger()` for `express`
+
+  ```ts
+  import express from "express";
+  import Logger from "Logger";
+  import { useExpressLogger } from "Logger";
+
+  const app = express();
+
+  const logger = new Logger();
+
+  app.use(
+    useExpressLogger(
+      logger /* Your logger here */,
+      logger.listSetting() /* Your setting of the logger */
+    )
+  );
+  ```
+
+- `levelLog` is now a `number` not an `Array`
+
+  ```ts
+  protected levelLog: IOLevelLog = IOLevelLog.NONE;
+  ```
+
+- Added custom level log for `prefix()` logging method
+
+  ```ts
+  logger.prefix({
+    prefix: "Prefix" /* <Your Prefix> */,
+    color: "red" /* <Your Color> */,
+    levelLog: 100 /* <Your Level> */,
+  });
+  ```
+
 ## 1.6.9
 
 - Small fixed, increase perfomance